refactor: extract open script Tabs component
